requirements.txt
glfw
numpy
platformdirs
PyGLM
PyOpenGL
pytools
typing_extensions
